add icons and do some good refactoring
