+++ package.json
@@ -29,6 +29,7 @@
     "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.3.0",
     "swr": "^2.2.4"
   },
   "devDependencies": {
